backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
gunicorn==21.2.0

